Vs in company

+++ LeetCode/MergeSortedArray.cpp
@@ -6,7 +6,7 @@ class SolutionMergeSortedArray
 {
 public:
 	void merge(vector<int>& nums1, int m, vector<int>& nums2, int n) {
-		for (int i = m; i < nums1.size(); i++)
+		for (int i = m; i < (int)nums1.size(); i++)
 			nums1[i] = nums2[i - m];
 		std::sort(nums1.begin(), nums1.end());
 	}
